feat: merge userService into accountService
